creditbank-backend/creditbank-backend/requirements.txt
Django==4.2.10
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
whitenoise==6.6.0
python-decouple==3.8
PyPI
